Set up a staging environment
This uses a different CF/S3 config: instead of enabling S3 website, this uses an origin policy to serve from "regular" S3. This has pros and cons. The S3 bucket can be kept private, but POST requests are not supported. To make navigator.sendBeacon work, I added a CF function that serves those requests with a simple 200 OK.